Added visualization for clustering
Generates a histogram for each feature in X which shows the distribution of values in each cluster. Generates a scatter plot of the X by reducing it to 2D using principle component analysis (PCA).

You can import this file and call visualize(X, Y, k) or you can run as a script (it currently uses the Boston data set).